Updated FavoriteRecipes Schema and Controllers
