+++ utils/mockRecipes.js
@@ -1,0 +1,473 @@
+const mockRecipes = [
+  {
+    id: 652332,
+    image: "https://img.spoonacular.com/recipes/652332-556x370.jpg",
+    imageType: "jpg",
+    title: "Monastery soup",
+    readyInMinutes: 45,
+    servings: 4,
+    sourceUrl: "https://www.foodista.com/recipe/Y67R5WMP/monastery-soup",
+    vegetarian: true,
+    vegan: true,
+    glutenFree: true,
+    dairyFree: true,
+    veryHealthy: false,
+    cheap: false,
+    veryPopular: false,
+    sustainable: false,
+    lowFodmap: false,
+    weightWatcherSmartPoints: 8,
+    gaps: "no",
+    preparationMinutes: null,
+    cookingMinutes: null,
+    aggregateLikes: 7,
+    healthScore: 39.0,
+    creditsText: "Foodista.com – The Cooking Encyclopedia Everyone Can Edit",
+    license: "CC BY 3.0",
+    sourceName: "Foodista",
+    pricePerServing: 156.78,
+    extendedIngredients: [
+      {
+        id: 11352,
+        aisle: "Produce",
+        image: "potatoes-yukon-gold.png",
+        consistency: "SOLID",
+        name: "potatoes",
+        nameClean: "potatoes",
+        original: "2 large potatoes, chopped into 1cm cubes",
+        originalName: "potatoes, chopped into 1cm cubes",
+        amount: 2.0,
+        unit: "large",
+        meta: ["chopped"],
+        measures: {
+          us: {
+            amount: 2.0,
+            unitShort: "large",
+            unitLong: "larges",
+          },
+          metric: {
+            amount: 2.0,
+            unitShort: "large",
+            unitLong: "larges",
+          },
+        },
+      },
+      {
+        id: 11282,
+        aisle: "Produce",
+        image: "brown-onion.png",
+        consistency: "SOLID",
+        name: "onion",
+        nameClean: "onion",
+        original: "1 large onion, chopped",
+        originalName: "onion, chopped",
+        amount: 1.0,
+        unit: "large",
+        meta: ["chopped"],
+        measures: {
+          us: {
+            amount: 1.0,
+            unitShort: "large",
+            unitLong: "large",
+          },
+          metric: {
+            amount: 1.0,
+            unitShort: "large",
+            unitLong: "large",
+          },
+        },
+      },
+      {
+        id: 11215,
+        aisle: "Produce",
+        image: "garlic.png",
+        consistency: "SOLID",
+        name: "garlic",
+        nameClean: "garlic",
+        original: "2 cloves garlic, chopped",
+        originalName: "garlic, chopped",
+        amount: 2.0,
+        unit: "cloves",
+        meta: ["chopped"],
+        measures: {
+          us: {
+            amount: 2.0,
+            unitShort: "cloves",
+            unitLong: "cloves",
+          },
+          metric: {
+            amount: 2.0,
+            unitShort: "cloves",
+            unitLong: "cloves",
+          },
+        },
+      },
+      {
+        id: 4584,
+        aisle: "Oil, Vinegar, Salad Dressing",
+        image: "vegetable-oil.jpg",
+        consistency: "LIQUID",
+        name: "sunflower oil",
+        nameClean: "sunflower oil",
+        original: "2 Tbs sunflower oil",
+        originalName: "sunflower oil",
+        amount: 2.0,
+        unit: "Tbs",
+        meta: [],
+        measures: {
+          us: {
+            amount: 2.0,
+            unitShort: "Tbs",
+            unitLong: "Tbs",
+          },
+          metric: {
+            amount: 2.0,
+            unitShort: "Tbs",
+            unitLong: "Tbs",
+          },
+        },
+      },
+      {
+        id: 11124,
+        aisle: "Produce",
+        image: "sliced-carrot.png",
+        consistency: "SOLID",
+        name: "carrots",
+        nameClean: "carrots",
+        original: "2 large carrots, peeled and chopped",
+        originalName: "carrots, peeled and chopped",
+        amount: 2.0,
+        unit: "large",
+        meta: ["peeled", "chopped"],
+        measures: {
+          us: {
+            amount: 2.0,
+            unitShort: "large",
+            unitLong: "larges",
+          },
+          metric: {
+            amount: 2.0,
+            unitShort: "large",
+            unitLong: "larges",
+          },
+        },
+      },
+      {
+        id: 11246,
+        aisle: "Produce",
+        image: "leeks.jpg",
+        consistency: "SOLID",
+        name: "leek",
+        nameClean: "leek",
+        original: "1 leek, washed thoroughly, sliced",
+        originalName: "leek, washed thoroughly, sliced",
+        amount: 1.0,
+        unit: "",
+        meta: ["washed", "sliced"],
+        measures: {
+          us: {
+            amount: 1.0,
+            unitShort: "",
+            unitLong: "",
+          },
+          metric: {
+            amount: 1.0,
+            unitShort: "",
+            unitLong: "",
+          },
+        },
+      },
+      {
+        id: 11143,
+        aisle: "Produce",
+        image: "celery.jpg",
+        consistency: "SOLID",
+        name: "celery",
+        nameClean: "celery",
+        original: "2 sticks celery, sliced",
+        originalName: "celery, sliced",
+        amount: 2.0,
+        unit: "sticks",
+        meta: ["sliced"],
+        measures: {
+          us: {
+            amount: 2.0,
+            unitShort: "sticks",
+            unitLong: "sticks",
+          },
+          metric: {
+            amount: 2.0,
+            unitShort: "sticks",
+            unitLong: "sticks",
+          },
+        },
+      },
+      {
+        id: 16053,
+        aisle: "Pasta and Rice",
+        image: "broad-beans.jpg",
+        consistency: "SOLID",
+        name: "broad beans",
+        nameClean: "broad beans",
+        original: "100g broad beans",
+        originalName: "broad beans",
+        amount: 100.0,
+        unit: "g",
+        meta: [],
+        measures: {
+          us: {
+            amount: 3.527,
+            unitShort: "oz",
+            unitLong: "ounces",
+          },
+          metric: {
+            amount: 100.0,
+            unitShort: "g",
+            unitLong: "grams",
+          },
+        },
+      },
+      {
+        id: 6615,
+        aisle: "Canned and Jarred",
+        image: "chicken-broth.png",
+        consistency: "LIQUID",
+        name: "vegetable stock",
+        nameClean: "vegetable stock",
+        original: "1 l vegetable stock",
+        originalName: "vegetable stock",
+        amount: 1.0,
+        unit: "l",
+        meta: [],
+        measures: {
+          us: {
+            amount: 1.046,
+            unitShort: "qt",
+            unitLong: "quarts",
+          },
+          metric: {
+            amount: 1.0,
+            unitShort: "l",
+            unitLong: "liter",
+          },
+        },
+      },
+      {
+        id: 10511297,
+        aisle: "Produce",
+        image: "parsley.jpg",
+        consistency: "SOLID",
+        name: "parsley",
+        nameClean: "parsley",
+        original: "Handful fresh parsley, chopped",
+        originalName: "fresh parsley, chopped",
+        amount: 1.0,
+        unit: "Handful",
+        meta: ["fresh", "chopped"],
+        measures: {
+          us: {
+            amount: 1.0,
+            unitShort: "Handful",
+            unitLong: "Handful",
+          },
+          metric: {
+            amount: 1.0,
+            unitShort: "Handful",
+            unitLong: "Handful",
+          },
+        },
+      },
+      {
+        id: 1102047,
+        aisle: "Spices and Seasonings",
+        image: "salt-and-pepper.jpg",
+        consistency: "SOLID",
+        name: "salt and pepper",
+        nameClean: "salt and pepper",
+        original: "Salt and pepper",
+        originalName: "Salt and pepper",
+        amount: 4.0,
+        unit: "servings",
+        meta: [],
+        measures: {
+          us: {
+            amount: 4.0,
+            unitShort: "servings",
+            unitLong: "servings",
+          },
+          metric: {
+            amount: 4.0,
+            unitShort: "servings",
+            unitLong: "servings",
+          },
+        },
+      },
+    ],
+    summary:
+      'Monastery soup is a <b>gluten free, dairy free, lacto ovo vegetarian, and vegan</b> recipe with 4 servings. One serving contains <b>293 calories</b>, <b>7g of protein</b>, and <b>8g of fat</b>. For <b>$1.57 per serving</b>, this recipe <b>covers 21%</b> of your daily requirements of vitamins and minerals. 7 people were glad they tried this recipe. It works best as a hor d\'oeuvre, and is done in approximately <b>45 minutes</b>. It will be a hit at your <b>Autumn</b> event. This recipe from Foodista requires parsley, carrots, broad beans, and vegetable stock. With a spoonacular <b>score of 89%</b>, this dish is super. Users who liked this recipe also liked <a href="https://spoonacular.com/recipes/monastery-soup-1412001">Monastery soup</a>, <a href="https://spoonacular.com/recipes/a-quick-and-easy-soup-miso-soup-with-soba-noodles-or-mung-bean-31003">A Quick And Easy Soup {miso Soup With Soba Noodles Or Mung Bean</a>, and <a href="https://spoonacular.com/recipes/pea-soup-with-lettuce-and-mint-aka-clean-out-the-fridge-soup-569347">Pea soup with lettuce and mint (aka: clean out the fridge soup!)</a>.',
+    cuisines: [],
+    dishTypes: [
+      "antipasti",
+      "soup",
+      "starter",
+      "snack",
+      "appetizer",
+      "antipasto",
+      "hor d'oeuvre",
+    ],
+    diets: ["gluten free", "dairy free", "lacto ovo vegetarian", "vegan"],
+    occasions: ["fall", "winter"],
+    instructions:
+      "1. Heat the oil in a large pan, add the potato, onion and garlic and saut, covered, for 10 minutes.\n2. Add the carrots, leek  and celery and saut for 5 minutes.\n3. Add broad beans and vegetable stock. Cover and simmer gently for 10 minutes, until the potatoes and broad beans are soft.\n4. Add parsley and salt  and pepper if needed.\nServe with crusty bread.",
+    analyzedInstructions: [
+      {
+        name: "",
+        steps: [
+          {
+            number: 1,
+            step: "Heat the oil in a large pan, add the potato, onion and garlic and saut, covered, for 10 minutes.",
+            ingredients: [
+              {
+                id: 11215,
+                name: "garlic",
+                localizedName: "garlic",
+                image: "garlic.png",
+              },
+              {
+                id: 11352,
+                name: "potato",
+                localizedName: "potato",
+                image: "potatoes-yukon-gold.png",
+              },
+              {
+                id: 11282,
+                name: "onion",
+                localizedName: "onion",
+                image: "brown-onion.png",
+              },
+              {
+                id: 4582,
+                name: "cooking oil",
+                localizedName: "cooking oil",
+                image: "vegetable-oil.jpg",
+              },
+            ],
+            equipment: [
+              {
+                id: 404645,
+                name: "frying pan",
+                localizedName: "frying pan",
+                image: "https://spoonacular.com/cdn/equipment_100x100/pan.png",
+              },
+            ],
+            length: {
+              number: 10,
+              unit: "minutes",
+            },
+          },
+          {
+            number: 2,
+            step: "Add the carrots, leek  and celery and saut for 5 minutes.",
+            ingredients: [
+              {
+                id: 11124,
+                name: "carrot",
+                localizedName: "carrot",
+                image: "sliced-carrot.png",
+              },
+              {
+                id: 11143,
+                name: "celery",
+                localizedName: "celery",
+                image: "celery.jpg",
+              },
+              {
+                id: 11246,
+                name: "leek",
+                localizedName: "leek",
+                image: "leeks.jpg",
+              },
+            ],
+            equipment: [],
+            length: {
+              number: 5,
+              unit: "minutes",
+            },
+          },
+          {
+            number: 3,
+            step: "Add broad beans and vegetable stock. Cover and simmer gently for 10 minutes, until the potatoes and broad beans are soft.",
+            ingredients: [
+              {
+                id: 6615,
+                name: "vegetable stock",
+                localizedName: "vegetable stock",
+                image: "chicken-broth.png",
+              },
+              {
+                id: 16053,
+                name: "fava beans",
+                localizedName: "fava beans",
+                image: "broad-beans.jpg",
+              },
+              {
+                id: 11352,
+                name: "potato",
+                localizedName: "potato",
+                image: "potatoes-yukon-gold.png",
+              },
+            ],
+            equipment: [],
+            length: {
+              number: 10,
+              unit: "minutes",
+            },
+          },
+          {
+            number: 4,
+            step: "Add parsley and salt  and pepper if needed.",
+            ingredients: [
+              {
+                id: 11297,
+                name: "parsley",
+                localizedName: "parsley",
+                image: "parsley.jpg",
+              },
+              {
+                id: 1002030,
+                name: "pepper",
+                localizedName: "pepper",
+                image: "pepper.jpg",
+              },
+              {
+                id: 2047,
+                name: "salt",
+                localizedName: "salt",
+                image: "salt.jpg",
+              },
+            ],
+            equipment: [],
+          },
+          {
+            number: 5,
+            step: "Serve with crusty bread.",
+            ingredients: [
+              {
+                id: 10018029,
+                name: "crusty bread",
+                localizedName: "crusty bread",
+                image: "crusty-bread.jpg",
+              },
+            ],
+            equipment: [],
+          },
+        ],
+      },
+    ],
+    originalId: null,
+    spoonacularScore: 88.15331268310547,
+    spoonacularSourceUrl: "https://spoonacular.com/monastery-soup-652332",
+  },
+];
+
+module.exports = { mockRecipes };
